+++ GoInvest/Modules/Sources/QuoteListModel/QuoteListModel.swift
